Re-structure components.
