+++ assets/js/modules/adsense/components/settings/AdBlockingRecoveryApp/Content.js
@@ -1,0 +1,60 @@
+/**
+ * Site Kit by Google, Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * External dependencies
+ */
+import PropTypes from 'prop-types';
+
+/**
+ * Internal dependencies
+ */
+import { Cell, Grid, Row } from '../../../../../material-components';
+import AdBlockingSetupSVG from '../../../../../../svg/graphics/ad-blocking-recovery-setup.svg';
+import {
+	BREAKPOINT_SMALL,
+	useBreakpoint,
+} from '../../../../../hooks/useBreakpoint';
+
+export default function Content( { children } ) {
+	const breakpoint = useBreakpoint();
+
+	const isTabletWidthOrLarger = breakpoint !== BREAKPOINT_SMALL;
+
+	return (
+		<Grid className="googlesitekit-ad-blocking-recovery__content">
+			<Row>
+				<Cell mdSize={ 6 } lgSize={ 8 }>
+					{ children }
+				</Cell>
+
+				{ isTabletWidthOrLarger && (
+					<Cell
+						className="googlesitekit-ad-blocking-recovery__hero-graphic"
+						mdSize={ 2 }
+						lgSize={ 4 }
+					>
+						<AdBlockingSetupSVG />
+					</Cell>
+				) }
+			</Row>
+		</Grid>
+	);
+}
+
+Content.propTypes = {
+	children: PropTypes.node,
+};
